Epic filters

+++ app/models.py
@@ -70,7 +70,7 @@ class Recipe(models.Model):
     username_id = models.TextField()
     creation_date = models.DateField(default=timezone.now)
     # Rating fileds
-    rating_average = models.DecimalField(max_digits=5, decimal_places=1, default=0.0)
+    rating_average = models.FloatField(default=0.0)
     rating_amount = models.IntegerField(default=0)
     rating_list = models.JSONField(default=dict)
